+++ app/src/org/commcare/connect/ConnectConstants.java
@@ -7,6 +7,8 @@ package org.commcare.connect;
  */
 public class ConnectConstants {
     public static final int ConnectIdTaskIdOffset = 1000;
+
+    public static int CREDENTIAL_PICKER_REQUEST = 2000;
     public static final String METHOD = "METHOD";
     public static final String CREATE = "CREATE";
     public static final String USERNAME = "USERNAME";
